requirements.txt
requests
onvif_zeep

pyyaml
fastapi
uvicorn
urllib3<2
requests_unixsocket

opencv-python
numpy
